requirements.txt
Flask==2.3.3
gunicorn==21.2.0
Werkzeug==2.3.7
itsdangerous==2.1.2
Jinja2==3.1.2
SQLAlchemy==1.4.49
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
flask-login==0.6.3
flask-cors==4.0.0
flask-jwt-extended==4.6.0
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.0
requests==2.31.0
